Input / Output Configuration through postcss.config.js
Tailwind-rn currently requires styles.json and tailwind.config.js to be in the root project directory.

Files names and path cannot be modified.  This feature aims to resolve this issue by relying on the paths outlined in postcss.config.js to provide a consistent configuration for the cli functionality, and routine build operations.

For example:

```
module.exports = {
  plugins: {
    tailwindcss: {
      config: 'app/assets/styles/tailwind.config.js',
      styles: 'app/assets/styles/styles.json'
    },
    autoprefixer: {},
  }
}
```